+++ .idea/misc.xml
@@ -8,10 +8,12 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/SDKs/ConciergeSample/app/src/main/res/layout/activity_banners_concierge.xml" value="0.3453125" />
         <entry key="..\:/SDKs/ConciergeSample/app/src/main/res/layout/activity_demo_appcompat_activity.xml" value="0.3015625" />
         <entry key="..\:/SDKs/ConciergeSample/app/src/main/res/layout/activity_main.xml" value="0.3015625" />
         <entry key="..\:/SDKs/ConciergeSample/app/src/main/res/layout/fragment_account.xml" value="0.3015625" />
         <entry key="..\:/SDKs/ConciergeSample/app/src/main/res/layout/fragment_login.xml" value="0.33" />
+        <entry key="..\:/SDKs/ConciergeSample/app/src/main/res/layout/fullscreen_concierge.xml" value="0.31666666666666665" />
         <entry key="..\:/SDKs/ConciergeSample/app/src/main/res/layout/layout_conciergeoptions_select.xml" value="0.19270833333333334" />
         <entry key="..\:/SDKs/ConciergeSample/app/src/main/res/layout/layout_viewholder1.xml" value="0.39947916666666666" />
         <entry key="..\:/SDKs/ConciergeSample/app/src/main/res/layout/layout_viewholder2.xml" value="0.39947916666666666" />
