+++ app/build.gradle
@@ -37,7 +37,7 @@ dependencies {
     implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
     implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
     implementation 'com.google.android.material:material:1.6.0'
-    implementation 'com.flybits.android:concierge:4.1.2'
+    implementation 'com.flybits.android:concierge:4.2.0'
     implementation 'com.google.firebase:firebase-core:18.0.2'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
