+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,16 @@
             android:theme="@style/AppTheme"
             tools:ignore="GoogleAppIndexingWarning">
         <activity
-                android:name=".activity.Fullscreen_Concierge"
+                android:name=".activity.BannersConcierge"
+                android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        <activity
+                android:name=".activity.FullscreenConcierge"
                 android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
